Removing files for refactor
Since the existing structure doesn't allow changes to be pushed to the
client, and knockout.js doesn't seem to support what I want.  So I've
deleted the main package as it exists currently so it can be rebuild using
a better structure, and going to in the future use just jquery.