Remove knockout from html and javascript
Removed the instances where knockout was used in the html and javascript
files.  Additionally, the page is no longer going to directly load the
config file, and instead get it over the websocket.  The remaining code
only connects to the websocket and has the send functionality which should
remain relatively the same.