New subscriptions get current status as events
Previously, as soon as vrpn was ready for events, the initial status of
the devices was broadcasted over events.  However, websockets will want
the current status when the connect.  Instead of storing that seperately
for websockets, and telling a connected websocket the status from that,
manager now gets the initial status from the config for each device.
Keeping track of the current status, whenever a new subscription happens
it gives the current status as a series of events to that subscription.
This handles the vrpn case more elegantly and solves the problem of any
subscription wanting the starting state.