Pass device info and events to web page
All the data is passed via a json object, which is immediately parsed into
an object on the web page.  The info includes what type of message it is
(a new button, slider, etc. or an update to an item) and the payload of
that data.  The payload for the events has a flag for whether the event
originated on the web page which will be used to prevent race conditions.